+++ components/frontend/src/report/Report.jsx
@@ -17,6 +17,7 @@ import { SubjectsButtonRow } from "../subject/SubjectsButtonRow"
 import { getReportTags } from "../utils"
 import { CommentSegment } from "../widgets/CommentSegment"
 import { WarningMessage } from "../widgets/WarningMessage"
+import { reverseSortMeasurements } from "./report_utils"
 import { ReportDashboard } from "./ReportDashboard"
 import { ReportTitle } from "./ReportTitle"
 
@@ -46,13 +47,13 @@ export function Report({
             </WarningMessage>
         )
     }
-    // Sort measurements in reverse order so that if there multiple measurements on a day, we find the most recent one:
-    const reversedMeasurements = measurements.slice().sort((m1, m2) => (m1.start < m2.start ? 1 : -1))
     return (
         <div id="dashboard">
             <PageHeader lastUpdate={lastUpdate} report={report} reportDate={reportDate} />
             <Paper elevation={5} sx={{ marginTop: "20px" }}>
                 <ReportTitle
+                    dates={dates}
+                    measurements={measurements}
                     openReportsOverview={openReportsOverview}
                     reload={reload}
                     report={report}
@@ -62,7 +63,7 @@ export function Report({
                 <Divider sx={{ padding: "0px" }} />
                 <ReportDashboard
                     dates={dates}
-                    measurements={reversedMeasurements}
+                    measurements={reverseSortMeasurements(measurements)}
                     onClick={(e, s) => navigateToSubject(e, s)}
                     onClickTag={(tag) => {
                         // If there are hidden tags (hiddenTags.length > 0), show the hidden tags.