+++ components/frontend/src/subject/SubjectTable.jsx
@@ -5,6 +5,7 @@ import { array, func, object, string } from "prop-types"
 import { useContext } from "react"
 
 import { DataModelContext } from "../context/DataModel"
+import { reverseSortMeasurements } from "../report/report_utils"
 import {
     availabilityMessagePropType,
     datesPropType,
@@ -34,8 +35,6 @@ export function SubjectTable({
     subjectUuid,
 }) {
     const dataModel = useContext(DataModelContext)
-    // Sort measurements in reverse order so that if there multiple measurements on a day, we find the most recent one:
-    const reversedMeasurements = measurements.slice().sort((m1, m2) => (m1.start < m2.start ? 1 : -1))
     const columnsToHide = determineColumnsToHide(dataModel, measurements, metricEntries, dates.length, report, settings)
     return (
         <TableContainer sx={{ overflowX: "visible" }}>
@@ -57,7 +56,7 @@ export function SubjectTable({
                     report={report}
                     reportDate={reportDate}
                     reports={reports}
-                    reversedMeasurements={reversedMeasurements}
+                    reversedMeasurements={reverseSortMeasurements(measurements)}
                     settings={settings}
                     subjectUuid={subjectUuid}
                 />